Move syncmessage.js to lib and change references to it
