+++ lib/syncmessage.js
