+++ server/routes/index.js
@@ -5,7 +5,7 @@ var middleware = require( './middleware' ),
     FilerWebServer = require('../lib/filer-www.js'),
     Sync = require( '../lib/sync'),
     ws = require('ws'),
-    SyncMessage = require('../lib/syncmessage'),
+    SyncMessage = require('../../lib/syncmessage'),
     WebSocket = require('ws'),
     WebSocketServer = WebSocket.Server,
     websocketAuth = require('../lib/websocket-auth');
